Extract SplitKind enum to break entity import cycle
ScheduledTransactionSplit imported SplitKind from transaction-split.entity,
which transitively pulled in account.entity -> scheduled-transaction.entity
-> scheduled-transaction-split.entity. At module-load the cached
transaction-split module had not finished evaluating yet, so SplitKind was
undefined when the @Column default was applied:

  TypeError: Cannot read properties of undefined (reading 'CATEGORY')
    at scheduled-transaction-split.entity.js

Move the enum to its own file (transactions/entities/split-kind.enum.ts)
and update all importers so the entities no longer participate in the
cycle.
