+++ backend/src/scheduled-transactions/dto/create-scheduled-transaction-split.dto.ts
@@ -15,7 +15,7 @@ import { Type } from "class-transformer";
 import { ApiPropertyOptional } from "@nestjs/swagger";
 import { SanitizeHtml } from "../../common/decorators/sanitize-html.decorator";
 import { InvestmentSplitDto } from "../../transactions/dto/create-transaction-split.dto";
-import { SplitKind } from "../../transactions/entities/transaction-split.entity";
+import { SplitKind } from "../../transactions/entities/split-kind.enum";
 
 export class CreateScheduledTransactionSplitDto {
   @ApiPropertyOptional({ enum: SplitKind })
