+++ backend/src/scheduled-transactions/dto/post-scheduled-transaction.dto.ts
@@ -14,7 +14,7 @@ import {
 import { Type } from "class-transformer";
 import { SanitizeHtml } from "../../common/decorators/sanitize-html.decorator";
 import { InvestmentSplitDto } from "../../transactions/dto/create-transaction-split.dto";
-import { SplitKind } from "../../transactions/entities/transaction-split.entity";
+import { SplitKind } from "../../transactions/entities/split-kind.enum";
 
 class InlineSplitDto {
   @ApiPropertyOptional({ enum: SplitKind })
