+++ backend/src/transactions/entities/split-kind.enum.ts
@@ -1,0 +1,5 @@
+export enum SplitKind {
+  CATEGORY = "category",
+  TRANSFER = "transfer",
+  INVESTMENT = "investment",
+}
